+++ tests/test_toolbox.py
@@ -373,19 +373,31 @@ class SimpleFunctionTests(unittest.TestCase):
         """feq should return true when they are equal"""
         val1 = 1.1234
         val2 = 1.1235
-        self.assertTrue(tb.feq(val1, val2, 0.0001))
-        numpy.testing.assert_array_equal(
-            [False, True, False, False],
-            tb.feq([1., 2., 3., 4.],
-                   [1.25, 2.05, 2.2, 500.1],
-                   0.1)
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter('always', category=DeprecationWarning)
+            self.assertTrue(tb.feq(val1, val2, 0.0001))
+            numpy.testing.assert_array_equal(
+                [False, True, False, False],
+                tb.feq([1., 2., 3., 4.],
+                       [1.25, 2.05, 2.2, 500.1],
+                       0.1)
             )
+        self.assertEqual(2, len(w))
+        for this_w in w:
+            self.assertEqual(DeprecationWarning, this_w.category)
+            self.assertEqual(DeprecationWarning, this_w.category)
+            self.assertEqual('use numpy.isclose', str(this_w.message))
 
     def testfeq_notequal(self):
         """feq should return false when they are not equal"""
         val1 = 1.1234
         val2 = 1.1235
-        self.assertFalse(tb.feq(val1, val2, 0.000005))
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter('always', category=DeprecationWarning)
+            self.assertFalse(tb.feq(val1, val2, 0.000005))
+        self.assertEqual(1, len(w))
+        self.assertEqual(DeprecationWarning, w[0].category)
+        self.assertEqual('use numpy.isclose', str(w[0].message))
 
     def test_medAbsDev(self):
         """medAbsDev should return a known range for given random input"""