+++ tests/test_spectrogram.py
@@ -7,6 +7,7 @@ Copyright ©2010 Los Alamos National Security, LLC.
 """
 
 import unittest
+import warnings
 
 import numpy as np
 import datetime
@@ -14,7 +15,11 @@ import datetime
 import spacepy.datamodel as dm
 import spacepy.toolbox as tb
 
-from spacepy.plot.spectrogram import spectrogram
+# Switch to this import once the name clash is gone
+#from spacepy.plot.spectrogram import Spectrogram
+from spacepy.plot import Spectrogram
+import spacepy.plot
+
 
 __all__ = ['spectrogramTests']
 
@@ -34,22 +39,33 @@ class spectrogramTests(unittest.TestCase):
 
     def test_keywords(self):
         """there is some input checking"""
-        self.assertRaises(KeyError, spectrogram, self.data, variables=['bad'])
-        self.assertRaises(KeyError, spectrogram, self.data, bad_keyword=['bad'])
+        self.assertRaises(KeyError, Spectrogram, self.data, variables=['bad'])
+        self.assertRaises(KeyError, Spectrogram, self.data, bad_keyword=['bad'])
 
     def test_init_raise_len(self):
         """__init__ does some checking on data length"""
         self.data['zval'] = []
-        self.assertRaises(ValueError, spectrogram, self.data, variables=self.kwargs['variables'])
+        self.assertRaises(ValueError, Spectrogram, self.data, variables=self.kwargs['variables'])
         self.data['yval'] = []
-        self.assertRaises(ValueError, spectrogram, self.data, variables=self.kwargs['variables'])
+        self.assertRaises(ValueError, Spectrogram, self.data, variables=self.kwargs['variables'])
         self.data['xval'] = []
-        self.assertRaises(ValueError, spectrogram, self.data, variables=self.kwargs['variables'])
-        
+        self.assertRaises(ValueError, Spectrogram, self.data, variables=self.kwargs['variables'])
+
+    def test_deprecation(self):
+        """Deprecation warning on old name"""
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter('always', category=DeprecationWarning)
+            a = spacepy.plot.spectrogram(
+                self.data, variables=self.kwargs['variables'])
+        self.assertEqual(1, len(w))
+        self.assertEqual(w[0].category, DeprecationWarning)
+        self.assertEqual(
+            'Use spacepy.plot.Spectrogram (capitalized).',
+            str(w[0].message))        
 
     def test_defaults(self):
         """run it and check that defaults were set correctly"""
-        a = spectrogram(self.data, variables=self.kwargs['variables'], extended_out=False)
+        a = Spectrogram(self.data, variables=self.kwargs['variables'], extended_out=False)
         ans = {'bins': [dm.dmarray([0.00120857, 0.07751865, 0.15382872, 0.2301388, 0.30644887,
                                     0.38275895, 0.45906902, 0.5353791, 0.61168917, 0.68799925,
                                     0.76430932, 0.8406194, 0.91692947, 0.99323955]),
@@ -71,7 +87,7 @@ class spectrogramTests(unittest.TestCase):
 
     def test_defaults_extended(self):
         """run it and check that defaults were set correctly (extended_out)"""
-        a = spectrogram(self.data, variables=self.kwargs['variables'])
+        a = Spectrogram(self.data, variables=self.kwargs['variables'])
         ans = {'bins': [dm.dmarray([0.00120857, 0.07751865, 0.15382872, 0.2301388, 0.30644887,
                                     0.38275895, 0.45906902, 0.5353791, 0.61168917, 0.68799925,
                                     0.76430932, 0.8406194, 0.91692947, 0.99323955]),
@@ -96,7 +112,7 @@ class spectrogramTests(unittest.TestCase):
                             zval = dm.dmarray(np.arange(3)))
         xbins = np.arange(-0.5, 3.5, 2.0)
         ybins = np.arange(-0.5, 3.5, 2.0)
-        a = spectrogram(self.data, variables=self.kwargs['variables'], extended_out=True)
+        a = Spectrogram(self.data, variables=self.kwargs['variables'], extended_out=True)
         count = a['spectrogram']['count'][:].copy()
         sm = a['spectrogram']['sum'][:].copy()
         spect = a['spectrogram']['spectrogram'][:].copy()
@@ -120,7 +136,7 @@ class spectrogramDateTests(unittest.TestCase):
 
     def test_defaults(self):
         """run it and check that defaults were set correctly"""
-        a = spectrogram(self.data, variables=self.kwargs['variables'])
+        a = Spectrogram(self.data, variables=self.kwargs['variables'])
         ans = {'bins': [dm.dmarray([730120.0, 730135.30769231, 730150.61538462,
                                     730165.92307692, 730181.23076923, 730196.53846154,
                                     730211.84615385, 730227.15384615, 730242.46153846,