+++ Doc/source/conf.py
@@ -62,7 +62,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'SpacePy'
-copyright = u'2011-2019, The SpacePy Team'
+copyright = u'2011-2020, The SpacePy Team'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -125,8 +125,8 @@ html_style = 'default.css'
 
 # The name for this set of Sphinx documents.  If None, it defaults to
 # "<project> v<release> documentation".
-html_title = "%s v%s Manual (DRAFT)" % (project, version)
-#html_title = "%s v%s Manual" % (project, version)
+#html_title = "%s v%s Manual (DRAFT)" % (project, version)
+html_title = "%s v%s Manual" % (project, version)
 
 # A shorter title for the navigation bar.  Default is the same as html_title.
 #html_short_title = None